Move and comment packages to follow new structure
