+++ fe2-android/app/src/androidTest/java/com/github/dedis/student20_pop/qrcode/CameraPermissionFragmentTest.java
@@ -1,0 +1,23 @@
+package com.github.dedis.student20_pop.qrcode;
+
+/** Class handling camera permission fragment tests */
+public class CameraPermissionFragmentTest {
+
+  //  @Rule
+  //  public ActivityScenarioRule<MainActivity> rule = new
+  // ActivityScenarioRule<>(MainActivity.class);
+  //
+  //  @Test
+  //  public void checkCorrectTabIsDisplayed() {
+  //    // Check that the permission is not granted
+  //    rule.getScenario()
+  //        .onActivity(
+  //            activity ->
+  //                assertThat(
+  //                    ActivityCompat.checkSelfPermission(activity, Manifest.permission.CAMERA),
+  //                    not(is(PackageManager.PERMISSION_GRANTED))));
+  //
+  //    onView(withId(R.id.tab_connect)).perform(click());
+  //    onView(withId(R.id.fragment_camera_perm)).check(matches(isDisplayed()));
+  //  }
+}
