+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/detail/fragments/event/creation/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Classes used to represent the different Event Creation UIs
+ *
+ * <p>The classes inside this package are fragments.
+ */
+package com.github.dedis.student20_pop.detail.fragments.event.creation;
