+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/home/HomeViewModel.java
@@ -19,8 +19,8 @@ import com.github.dedis.student20_pop.model.network.answer.Error;
 import com.github.dedis.student20_pop.model.network.answer.Result;
 import com.github.dedis.student20_pop.model.network.method.message.MessageGeneral;
 import com.github.dedis.student20_pop.model.network.method.message.data.lao.CreateLao;
-import com.github.dedis.student20_pop.ui.qrcode.CameraPermissionViewModel;
-import com.github.dedis.student20_pop.ui.qrcode.QRCodeScanningViewModel;
+import com.github.dedis.student20_pop.qrcode.CameraPermissionViewModel;
+import com.github.dedis.student20_pop.qrcode.QRCodeScanningViewModel;
 import com.github.dedis.student20_pop.utility.security.Keys;
 import com.google.android.gms.vision.barcode.Barcode;
 import com.google.crypto.tink.KeysetHandle;
