+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/home/fragments/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Classes used to represent the different UIs of the home view
+ */
+package com.github.dedis.student20_pop.home.fragments;
