+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/home/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Classes related to the UI and utility of the home view
+ *
+ * </p> The needed activities, fragments, adapters and listeners are inside this package.
+ * </p> One view model is present for the logic of the home view.
+ */
+package com.github.dedis.student20_pop.home;
