+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/qrcode/CameraPermissionViewModel.java
@@ -1,0 +1,6 @@
+package com.github.dedis.student20_pop.qrcode;
+
+public interface CameraPermissionViewModel {
+
+  void onPermissionGranted();
+}
