+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/qrcode/QRCodeScanningViewModel.java
@@ -1,0 +1,8 @@
+package com.github.dedis.student20_pop.qrcode;
+
+import com.github.dedis.student20_pop.utility.qrcode.QRCodeListener;
+
+public interface QRCodeScanningViewModel extends QRCodeListener {
+
+  int getScanDescription();
+}
