+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/qrcode/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Classes used to represent the different UIs and view models linked to QR code scanning
+ */
+package com.github.dedis.student20_pop.qrcode;
