--- fe2-android/app/src/main/java/com/github/dedis/student20_pop/ui/event/creation/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Classes used to represent the different Event Creation UIs
- *
- * <p>The classes inside this package are fragments.
- */
-package com.github.dedis.student20_pop.ui.event.creation;
