--- fe2-android/app/src/main/java/com/github/dedis/student20_pop/ui/qrcode/CameraPermissionViewModel.java
@@ -1,6 +1,0 @@
-package com.github.dedis.student20_pop.ui.qrcode;
-
-public interface CameraPermissionViewModel {
-
-  void onPermissionGranted();
-}
