--- fe2-android/app/src/main/java/com/github/dedis/student20_pop/ui/qrcode/QRCodeScanningViewModel.java
@@ -1,8 +1,0 @@
-package com.github.dedis.student20_pop.ui.qrcode;
-
-import com.github.dedis.student20_pop.utility.qrcode.QRCodeListener;
-
-public interface QRCodeScanningViewModel extends QRCodeListener {
-
-  int getScanDescription();
-}
