--- fe2-android/app/src/main/java/com/github/dedis/student20_pop/utility/ui/adapter/package-info.java
@@ -1,2 +1,0 @@
-/** Classes used to help the representation of list and expandable lists */
-package com.github.dedis.student20_pop.utility.ui.adapter;
