--- fe2-android/app/src/main/java/com/github/dedis/student20_pop/utility/ui/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Classes used to display and listen to the elements in relation to the different UIs.
- *
- * <p>The classes inside this package are adapters and listeners.
- */
-package com.github.dedis.student20_pop.utility.ui;
